fix: integração com o BD

+++ backend/database/migrations/2026_05_09_165230_create_alunos_table.php
@@ -6,26 +6,22 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('alunos', function (Blueprint $table) {
-            $table->id();
-            $table->string('nome');
-            $table->string('matricula')->unique();
-            $table->string('email')->nullable();
-            $table->string('telefone')->nullable();
-            $table->string('turma');
-            $table->foreignId('usuario_id')->nullable()->references('id')->on('usuarios')->onDelete('set null');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('alunos')) {
+            Schema::create('alunos', function (Blueprint $table) {
+                $table->id();
+                $table->string('nome');
+                $table->string('matricula')->unique();
+                $table->string('email')->nullable();
+                $table->string('telefone')->nullable();
+                $table->string('turma');
+                $table->foreignId('usuario_id')->nullable()->references('id')->on('usuarios')->onDelete('set null');
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('alunos');
