+++ backend/database/migrations/2026_05_09_165231_create_faltas_table.php
@@ -6,23 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('faltas', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('aluno_id')->references('id')->on('alunos')->onDelete('cascade');
-            $table->date('data');
-            $table->boolean('presente')->default(true);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('faltas')) {
+            Schema::create('faltas', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('aluno_id')->references('id')->on('alunos')->onDelete('cascade');
+                $table->date('data');
+                $table->boolean('presente')->default(true);
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('faltas');
