+++ backend/database/migrations/2026_05_09_165231_create_notas_table.php
@@ -6,24 +6,20 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::create('notas', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('aluno_id')->references('id')->on('alunos')->onDelete('cascade');
-            $table->foreignId('disciplina_id')->references('id')->on('disciplinas')->onDelete('cascade');
-            $table->decimal('valor_nota', 5, 2);
-            $table->integer('semestre')->default(1);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('notas')) {
+            Schema::create('notas', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('aluno_id')->references('id')->on('alunos')->onDelete('cascade');
+                $table->foreignId('disciplina_id')->references('id')->on('disciplinas')->onDelete('cascade');
+                $table->decimal('valor_nota', 5, 2);
+                $table->integer('semestre')->default(1);
+                $table->timestamps();
+            });
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::dropIfExists('notas');
